Fixes: Damage calculation, enchantments now binded and stored through ExtraData and not a dedicates list. Hacky, but MiNET..

+++ src/MiNET/MiNET/Items/Enchantments/EnchantmentType.cs
@@ -1,6 +1,6 @@
 ﻿namespace MiNET.Items.Enchantments
 {
-	public enum EnchantmentType : int
+	public enum EnchantmentType : short
 	{
 		Invalid = -1,
 		Protection = 0,
